+++ libqaul/src/api/models.rs
@@ -1,7 +1,6 @@
 //! Service API exchange models
 
 use identity::Identity;
-use std::collections::BTreeMap;
 
 /// Convenience type for API functions
 pub type QaulResult<T> = Result<T, QaulError>;
@@ -58,27 +57,4 @@ pub struct Message {
     recipient: Identity,
     payload: Vec<u8>,
     signature: SigTrust,
-}
-
-/// A public representation of user metadata
-///
-/// Apart from the user `id`, all fields are optional
-/// and should not be assumed set. This struct is used
-/// for both the local user (identified by `UserAuth`)
-/// as well as remote users from the contacts book.
-pub struct User {
-    /// A users network ID
-    pub id: Identity,
-    /// A human readable display-name (like @foobar)
-    pub display_name: Option<String>,
-    /// A human's preferred call-signed ("Friends call be foo")
-    pub real_name: Option<String>,
-    /// A key-value list of things the user deems interesting
-    /// about themselves. This could be stuff like "gender",
-    /// "preferred languages" or whatever.
-    pub bio: BTreeMap<String, String>,
-    /// The set of services this user runs (should never be empty!)
-    pub services: Vec<String>,
-    /// A users profile picture (some people like selfies)
-    pub avatar: Option<Vec<u8>>,
 }
